Three improvements, including adjacency measure, gamma colours, computing of id_a and id_c
Improved setting of gammaColour_byid to handle arbitrary ranges of
gamma values (important for the gamma_noise work)

Computes an adjacency vector for each TC ID

As part of the above, now has code to determine id_c and id_a. This
wasn't previously necessary for id_c, because usually it was computed
and saved by the simulation, but now it can be reconstructed even when
it wasn't computed by the c++ program.